+++ src/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import {Routes, Route} from "react-router-dom";
 import Main from "./Pages/Main";
-import Login from "./Pages/Login";
-import SignUp from "./Pages/SignUp";
-import Community from "./Pages/Community";
-import LoginRedirect from "./Pages/LoginRedirect";
+import Login from "./Pages/Start/Login";
+import SignUp from "./Pages/Start/SignUp";
+import Community from "./Pages/Board/Community";
+import LoginRedirect from "./Pages/Start/LoginRedirect";
 
 const App = () => {
   return (
@@ -13,7 +13,7 @@ const App = () => {
       <Route path="/login" element={<Login/>} />
       <Route path="/signup" element={<SignUp/>} />
       <Route path="/community" element={<Community/>} />
-      <Route path="/oauth/redirect" element={<LoginRedirect/>} />
+      <Route path="/oauth/kakao" element={<LoginRedirect/>} />
     </Routes>
   );
 }
